implemented showing search results;refactored code
